requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==3.3
mysql==0.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.1.1
pandas==1.4.2
persiantools==3.0.0
Pillow==8.4.0
protobuf==3.20.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
PyMySQL==1.0.2
pyodbc==4.0.32
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
